add function to generate artifacts
