+++ src/commands/generate.ts
@@ -3,6 +3,7 @@ import { findServerEntryPoint } from '../utils'
 import { spawnSync } from 'child_process'
 import * as path from 'path'
 import * as fs from 'fs'
+import { generateArtifacts } from '../utils/artifact-generation'
 
 export class Generate extends Command {
   static description = 'Generate the artifacts'
@@ -17,27 +18,12 @@ export class Generate extends Command {
 
   async run() {
     const { args, flags } = this.parse(Generate)
-    const entryPoint = flags.entrypoint
-      ? path.join(process.cwd(), flags.entrypoint)
-      : findServerEntryPoint()
+    const { error } = generateArtifacts(flags.entrypoint)
 
-    if (!fs.existsSync(entryPoint)) {
-      this.error(
-        `🎃  Entry point "${path.relative(
-          process.cwd(),
-          entryPoint
-        )}" does not exist`,
-        { exit: 1 }
-      )
+    if (error) {
+      this.error(error, { exit: 1 })
     }
-
-    spawnSync('ts-node', [entryPoint], {
-      env: {
-        ...process.env,
-        PUMPKINS_SHOULD_EXIT_AFTER_GENERATE_ARTIFACTS: 'true',
-      },
-    })
-
+    
     this.log('🎃  Successfully generated the artifacts')
   }
 }
